# swagger.yaml
openapi: 3.0.0
info:
  title: 'API Livros'
  description: 'Documentação da API Livros'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080/api-livros'
    description: 'Servidor local'
paths:
...
  /usuario/foto:
    put:
      tags:
        - Usuário
      summary: 'Atualiza a foto de perfil do usuário autenticado'
      operationId: a76c0610241bf27126a5aaf1daaf901e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioFotoRequest'
      responses:
        '200':
          description: 'Foto atualizada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioFotoResponse'
        '400':
          description: 'URL inválida'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Erro interno'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /usuario:
    get:
      tags:
        - Usuário
      summary: 'Lista os dados do usuário autenticado'
      operationId: 1ee74c0a21f0b1d380b76ab89163a811
      responses:
        '200':
          description: 'Usuário encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioListResponse'
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Usuário não encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Usuário
      summary: 'Edita os dados do usuário autenticado'
      operationId: 9096c93a3a91a723d09c1fb0c1f9db71
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioUpdateRequest'
      responses:
        '200':
          description: 'Usuário atualizado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioUpdateResponse'
        '401':
          description: 'Não autenticado'
        '409':
          description: 'Email já em uso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Erro interno'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Usuário
      summary: 'Deleta o usuário autenticado'
      operationId: c22821ed743d1fe670a7ff3a999c1f5f
      responses:
        '200':
          description: 'Usuário deletado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDeleteResponse'
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Erro interno'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
components:
  schemas:
...
    ErrorResponse:
      properties:
        success:
          type: boolean
          example: false
        mensagem:
          type: string
          example: 'Erro ao processar requisição'
      type: object
...
    UsuarioFotoRequest:
      required:
        - url_foto
      properties:
        url_foto:
          type: string
          format: uri
          example: 'https://meusite.com/imagens/foto.jpg'
      type: object
    UsuarioUpdateRequest:
      properties:
        nome:
          type: string
          example: 'Luiz Felipe'
        email:
          type: string
          format: email
          example: luiz@email.com
      type: object
    UsuarioResource:
      properties:
        id_usuario:
          type: integer
          example: 1
        UUID:
          type: string
          example: e04230085f34fcdc518137ac826725
        nome:
          type: string
          example: 'Luiz Felipe'
        email:
          type: string
          format: email
          example: teste@email.com
        foto_perfil:
          type: string
          example: null
          nullable: true
      type: object
    UsuarioFotoResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          properties:
            mensagem:
              type: string
              example: 'Foto de perfil atualizada com sucesso'
            foto_perfil:
              type: string
              format: uri
              example: 'https://meusite.com/imagens/foto.jpg'
          type: object
      type: object
    UsuarioUpdateResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          properties:
            mensagem:
              type: string
              example: 'Usuario atualizado com sucesso'
            usuario:
              properties:
                id_usuario:
                  type: integer
                  example: 1
              type: object
          type: object
      type: object
    UsuarioDeleteResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          properties:
            mensagem:
              type: string
              example: 'Usuário deletado com sucesso'
          type: object
      type: object
    UsuarioListResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          properties:
            mensagem:
              type: string
              example: 'Usuário encontrado'
            usuario:
              $ref: '#/components/schemas/UsuarioResource'
          type: object
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
    userUuid:
      type: apiKey
      description: 'UUID do usuário'
      name: X-USER-UUID
      in: header
...